src/test/java/com/worldofbooks/listingsreport/output/FtpClientTest.java
package com.worldofbooks.listingsreport.output;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.mockftpserver.fake.FakeFtpServer;
import org.mockftpserver.fake.UserAccount;
import org.mockftpserver.fake.filesystem.DirectoryEntry;
import org.mockftpserver.fake.filesystem.FileEntry;
import org.mockftpserver.fake.filesystem.FileSystem;
import org.mockftpserver.fake.filesystem.UnixFakeFileSystem;

import java.io.IOException;
import java.net.URISyntaxException;
import static org.junit.Assert.assertTrue;

public class FtpClientTest {

    private FakeFtpServer fakeFtpServer;

    private FtpClient ftpClient;

    @Before
    public void setup() throws IOException {
        fakeFtpServer = new FakeFtpServer();
        fakeFtpServer.addUserAccount(new UserAccount("ftpUser", "ftpPassword", "/test"));

        FileSystem fileSystem = new UnixFakeFileSystem();
        fileSystem.add(new DirectoryEntry("/test"));
        fileSystem.add(new FileEntry("/test/ftpTest.txt", "test file content"));
        fakeFtpServer.setFileSystem(fileSystem);
        fakeFtpServer.setServerControlPort(0);

        fakeFtpServer.start();

        ftpClient = new FtpClient("localhost", fakeFtpServer.getServerControlPort(), "ftpUser", "ftpPassword");
        ftpClient.open();
    }

    @After
    public void teardown() throws IOException {
        ftpClient.close();
        fakeFtpServer.stop();
    }

    @Test
    public void sendToFtp() throws IOException {
        ftpClient.sendToFtp("ftpTest.txt", "/uploadTest.txt");
        assertTrue(fakeFtpServer.getFileSystem().exists("/uploadTest.txt"));
    }

}
